Fix run command tests
