+++ tests/conftest.py
@@ -1,4 +1,7 @@
 """Reusable fixtures."""
+import tempfile
+from typing import Iterator
+
 import pytest
 from typer.testing import CliRunner
 
@@ -6,3 +9,19 @@ from typer.testing import CliRunner
 @pytest.fixture
 def runner():
     return CliRunner()
+
+
+@pytest.fixture
+def matcha_testing_directory() -> Iterator[str]:
+    """A fixture for creating and removing temporary test directory for storing and moving files.
+
+    Yields:
+        str: a path to temporary directory for storing and moving files from tests.
+    """
+    temp_dir = tempfile.TemporaryDirectory()
+
+    # tests are executed at this point
+    yield temp_dir.name
+
+    # delete temp folder
+    temp_dir.cleanup()
