+++ tests/test_cli/test_run.py
@@ -9,20 +9,23 @@ from matcha_ml.cli.cli import app
 BASE_DIR = os.path.dirname(os.path.abspath(__file__))
 
 
-@pytest.fixture(autouse=True)
-def create_and_teardown_run_py_file():
+@pytest.fixture
+def run_testing_directory(matcha_testing_directory: str):
     """A fixture for creating and removing a temporary run.py file.
 
     To test whether run.py is executed when the `matcha run` is used with no other commands.
+
+    Args:
+        matcha_testing_directory (str): temporary working directory
 
     Yields:
         run.py: a run.py temporarily containing 1 line of python code which prints "This is the run.py file".
     """
 
-    with open("run.py", "w") as f:
+    with open(os.path.join(matcha_testing_directory, "run.py"), "w") as f:
         f.write("print('This is the run.py file')")
 
-    yield  # tests are executed at this point
+    yield matcha_testing_directory  # tests are executed at this point
 
     # then the run.py file is remove
     os.remove("run.py")
@@ -40,7 +43,7 @@ def test_cli_run_command(runner):
     assert "Run command." in result.stdout
 
 
-def test_cli_train_command(runner):
+def test_cli_train_command(runner, matcha_testing_directory):
     """Test cli for run command."""
     # Invoke run command
     result = runner.invoke(app, ["run", "train", "--help"])
@@ -52,9 +55,10 @@ def test_cli_train_command(runner):
     assert "Run train subcommand." in result.stdout
 
 
-def test_cli_default_callback(runner):
+def test_cli_default_callback(runner, run_testing_directory: str):
     """Test cli for run command."""
     # Invoke run command with no option passed
+    os.chdir(run_testing_directory)
     result = runner.invoke(app, ["run"])
 
     # Exit code 0 means there was no error
